+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 
 ### Enhancements
 
+- Gating and Scheduling
 
 ## 0.16.0 (2021-11-19)